refactor: refactor code to receive botId by prop drilling. Make sure that clear chat will display default message without calling API again
